+++ ca/ca_test.go
@@ -24,7 +24,6 @@ import (
 	ctx509 "github.com/google/certificate-transparency-go/x509"
 	"github.com/jmhodges/clock"
 	"github.com/prometheus/client_golang/prometheus"
-	"golang.org/x/crypto/ocsp"
 
 	capb "github.com/letsencrypt/boulder/ca/proto"
 	"github.com/letsencrypt/boulder/cmd"
@@ -131,15 +130,17 @@ func mustRead(path string) []byte {
 
 type testCtx struct {
 	pa             core.PolicyAuthority
+	ocsp           *ocspImpl
 	certExpiry     time.Duration
 	certBackdate   time.Duration
 	serialPrefix   int
 	maxNames       int
-	ocspLifetime   time.Duration
 	boulderIssuers []*issuance.Issuer
 	keyPolicy      goodkey.KeyPolicy
 	fc             clock.FakeClock
 	stats          prometheus.Registerer
+	signatureCount *prometheus.CounterVec
+	signErrorCount *prometheus.CounterVec
 	logger         *blog.Mock
 }
 
@@ -241,18 +242,44 @@ func setup(t *testing.T) *testCtx {
 		AllowECDSANISTP256: true,
 		AllowECDSANISTP384: true,
 	}
+	signatureCount := prometheus.NewCounterVec(
+		prometheus.CounterOpts{
+			Name: "signatures",
+			Help: "Number of signatures",
+		},
+		[]string{"purpose", "issuer"})
+	signErrorCount := prometheus.NewCounterVec(prometheus.CounterOpts{
+		Name: "signature_errors",
+		Help: "A counter of signature errors labelled by error type",
+	}, []string{"type"})
+
+	ocsp, err := NewOCSPImpl(
+		&mockSA{},
+		boulderIssuers,
+		time.Hour,
+		0,
+		time.Second,
+		blog.NewMock(),
+		metrics.NoopRegisterer,
+		signatureCount,
+		signErrorCount,
+		fc,
+	)
+	test.AssertNotError(t, err, "Failed to create ocsp impl")
 
 	return &testCtx{
 		pa:             pa,
+		ocsp:           ocsp,
 		certExpiry:     8760 * time.Hour,
 		certBackdate:   time.Hour,
 		serialPrefix:   17,
 		maxNames:       2,
-		ocspLifetime:   time.Hour,
 		boulderIssuers: boulderIssuers,
 		keyPolicy:      keyPolicy,
 		fc:             fc,
 		stats:          metrics.NoopRegisterer,
+		signatureCount: signatureCount,
+		signErrorCount: signErrorCount,
 		logger:         blog.NewMock(),
 	}
 }
@@ -265,17 +292,17 @@ func TestFailNoSerialPrefix(t *testing.T) {
 		nil,
 		nil,
 		nil,
+		nil,
 		testCtx.certExpiry,
 		testCtx.certBackdate,
 		0,
 		testCtx.maxNames,
-		testCtx.ocspLifetime,
 		testCtx.keyPolicy,
 		nil,
-		0,
-		time.Second,
 		testCtx.logger,
 		testCtx.stats,
+		nil,
+		nil,
 		testCtx.fc)
 	test.AssertError(t, err, "CA should have failed with no SerialPrefix")
 }
@@ -362,19 +389,19 @@ func issueCertificateSubTestSetup(t *testing.T) (*certificateAuthorityImpl, *moc
 	ca, err := NewCertificateAuthorityImpl(
 		sa,
 		testCtx.pa,
+		testCtx.ocsp,
 		testCtx.boulderIssuers,
 		nil,
 		testCtx.certExpiry,
 		testCtx.certBackdate,
 		testCtx.serialPrefix,
 		testCtx.maxNames,
-		testCtx.ocspLifetime,
 		testCtx.keyPolicy,
 		nil,
-		0,
-		time.Second,
 		testCtx.logger,
 		testCtx.stats,
+		testCtx.signatureCount,
+		testCtx.signErrorCount,
 		testCtx.fc)
 	test.AssertNotError(t, err, "Failed to create CA")
 
@@ -410,19 +437,19 @@ func TestMultipleIssuers(t *testing.T) {
 	ca, err := NewCertificateAuthorityImpl(
 		sa,
 		testCtx.pa,
+		testCtx.ocsp,
 		testCtx.boulderIssuers,
 		nil,
 		testCtx.certExpiry,
 		testCtx.certBackdate,
 		testCtx.serialPrefix,
 		testCtx.maxNames,
-		testCtx.ocspLifetime,
 		testCtx.keyPolicy,
 		nil,
-		0,
-		time.Second,
 		testCtx.logger,
 		testCtx.stats,
+		testCtx.signatureCount,
+		testCtx.signErrorCount,
 		testCtx.fc)
 	test.AssertNotError(t, err, "Failed to remake CA")
 
@@ -475,89 +502,6 @@ func TestECDSAAllowList(t *testing.T) {
 	test.AssertByteEquals(t, cert.RawIssuer, caCert2.RawSubject)
 }
 
-func TestOCSP(t *testing.T) {
-	testCtx := setup(t)
-	sa := &mockSA{}
-	ca, err := NewCertificateAuthorityImpl(
-		sa,
-		testCtx.pa,
-		testCtx.boulderIssuers,
-		nil,
-		testCtx.certExpiry,
-		testCtx.certBackdate,
-		testCtx.serialPrefix,
-		testCtx.maxNames,
-		testCtx.ocspLifetime,
-		testCtx.keyPolicy,
-		nil,
-		0,
-		time.Second,
-		testCtx.logger,
-		testCtx.stats,
-		testCtx.fc)
-	test.AssertNotError(t, err, "Failed to create CA")
-
-	// Issue a certificate from the RSA issuer caCert, then check OCSP comes from the same issuer.
-	rsaIssuerID := ca.issuers.byAlg[x509.RSA].cert.ID()
-	rsaCertPB, err := ca.IssuePrecertificate(ctx, &capb.IssueCertificateRequest{Csr: CNandSANCSR, RegistrationID: arbitraryRegID})
-	test.AssertNotError(t, err, "Failed to issue certificate")
-	rsaCert, err := x509.ParseCertificate(rsaCertPB.DER)
-	test.AssertNotError(t, err, "Failed to parse rsaCert")
-	rsaOCSPPB, err := ca.GenerateOCSP(ctx, &capb.GenerateOCSPRequest{
-		Serial:   core.SerialToString(rsaCert.SerialNumber),
-		IssuerID: int64(rsaIssuerID),
-		Status:   string(core.OCSPStatusGood),
-	})
-	test.AssertNotError(t, err, "Failed to generate OCSP")
-	rsaOCSP, err := ocsp.ParseResponse(rsaOCSPPB.Response, caCert.Certificate)
-	test.AssertNotError(t, err, "Failed to parse / validate OCSP for rsaCert")
-	test.AssertEquals(t, rsaOCSP.Status, 0)
-	test.AssertEquals(t, rsaOCSP.RevocationReason, 0)
-	test.AssertEquals(t, rsaOCSP.SerialNumber.Cmp(rsaCert.SerialNumber), 0)
-
-	// Issue a certificate from the ECDSA issuer caCert2, then check OCSP comes from the same issuer.
-	ecdsaIssuerID := ca.issuers.byAlg[x509.ECDSA].cert.ID()
-	ecdsaCertPB, err := ca.IssuePrecertificate(ctx, &capb.IssueCertificateRequest{Csr: ECDSACSR, RegistrationID: arbitraryRegID})
-	test.AssertNotError(t, err, "Failed to issue certificate")
-	ecdsaCert, err := x509.ParseCertificate(ecdsaCertPB.DER)
-	test.AssertNotError(t, err, "Failed to parse ecdsaCert")
-	ecdsaOCSPPB, err := ca.GenerateOCSP(ctx, &capb.GenerateOCSPRequest{
-		Serial:   core.SerialToString(ecdsaCert.SerialNumber),
-		IssuerID: int64(ecdsaIssuerID),
-		Status:   string(core.OCSPStatusGood),
-	})
-	test.AssertNotError(t, err, "Failed to generate OCSP")
-	ecdsaOCSP, err := ocsp.ParseResponse(ecdsaOCSPPB.Response, caCert2.Certificate)
-	test.AssertNotError(t, err, "Failed to parse / validate OCSP for ecdsaCert")
-	test.AssertEquals(t, ecdsaOCSP.Status, 0)
-	test.AssertEquals(t, ecdsaOCSP.RevocationReason, 0)
-	test.AssertEquals(t, ecdsaOCSP.SerialNumber.Cmp(ecdsaCert.SerialNumber), 0)
-
-	// GenerateOCSP with a bad IssuerID should fail.
-	_, err = ca.GenerateOCSP(context.Background(), &capb.GenerateOCSPRequest{
-		Serial:   core.SerialToString(rsaCert.SerialNumber),
-		IssuerID: int64(666),
-		Status:   string(core.OCSPStatusGood),
-	})
-	test.AssertError(t, err, "GenerateOCSP didn't fail with invalid IssuerID")
-
-	// GenerateOCSP with a bad Serial should fail.
-	_, err = ca.GenerateOCSP(context.Background(), &capb.GenerateOCSPRequest{
-		Serial:   "BADDECAF",
-		IssuerID: int64(rsaIssuerID),
-		Status:   string(core.OCSPStatusGood),
-	})
-	test.AssertError(t, err, "GenerateOCSP didn't fail with invalid Serial")
-
-	// GenerateOCSP with a valid-but-nonexistent Serial should *not* fail.
-	_, err = ca.GenerateOCSP(context.Background(), &capb.GenerateOCSPRequest{
-		Serial:   "03DEADBEEFBADDECAFFADEFACECAFE30",
-		IssuerID: int64(rsaIssuerID),
-		Status:   string(core.OCSPStatusGood),
-	})
-	test.AssertNotError(t, err, "GenerateOCSP failed with fake-but-valid Serial")
-}
-
 func TestInvalidCSRs(t *testing.T) {
 	testCases := []struct {
 		name         string
@@ -607,19 +551,19 @@ func TestInvalidCSRs(t *testing.T) {
 		ca, err := NewCertificateAuthorityImpl(
 			sa,
 			testCtx.pa,
+			testCtx.ocsp,
 			testCtx.boulderIssuers,
 			nil,
 			testCtx.certExpiry,
 			testCtx.certBackdate,
 			testCtx.serialPrefix,
 			testCtx.maxNames,
-			testCtx.ocspLifetime,
 			testCtx.keyPolicy,
 			nil,
-			0,
-			time.Second,
 			testCtx.logger,
 			testCtx.stats,
+			testCtx.signatureCount,
+			testCtx.signErrorCount,
 			testCtx.fc)
 		test.AssertNotError(t, err, "Failed to create CA")
 
@@ -645,19 +589,19 @@ func TestRejectValidityTooLong(t *testing.T) {
 	ca, err := NewCertificateAuthorityImpl(
 		sa,
 		testCtx.pa,
+		testCtx.ocsp,
 		testCtx.boulderIssuers,
 		nil,
 		testCtx.certExpiry,
 		testCtx.certBackdate,
 		testCtx.serialPrefix,
 		testCtx.maxNames,
-		testCtx.ocspLifetime,
 		testCtx.keyPolicy,
 		nil,
-		0,
-		time.Second,
 		testCtx.logger,
 		testCtx.stats,
+		nil,
+		nil,
 		testCtx.fc)
 	test.AssertNotError(t, err, "Failed to create CA")
 
@@ -747,19 +691,19 @@ func TestIssueCertificateForPrecertificate(t *testing.T) {
 	ca, err := NewCertificateAuthorityImpl(
 		sa,
 		testCtx.pa,
+		testCtx.ocsp,
 		testCtx.boulderIssuers,
 		nil,
 		testCtx.certExpiry,
 		testCtx.certBackdate,
 		testCtx.serialPrefix,
 		testCtx.maxNames,
-		testCtx.ocspLifetime,
 		testCtx.keyPolicy,
 		nil,
-		0,
-		time.Second,
 		testCtx.logger,
 		testCtx.stats,
+		testCtx.signatureCount,
+		testCtx.signErrorCount,
 		testCtx.fc)
 	test.AssertNotError(t, err, "Failed to create CA")
 
@@ -854,19 +798,19 @@ func TestIssueCertificateForPrecertificateDuplicateSerial(t *testing.T) {
 	ca, err := NewCertificateAuthorityImpl(
 		sa,
 		testCtx.pa,
+		testCtx.ocsp,
 		testCtx.boulderIssuers,
 		nil,
 		testCtx.certExpiry,
 		testCtx.certBackdate,
 		testCtx.serialPrefix,
 		testCtx.maxNames,
-		testCtx.ocspLifetime,
 		testCtx.keyPolicy,
 		nil,
-		0,
-		time.Second,
 		testCtx.logger,
 		testCtx.stats,
+		testCtx.signatureCount,
+		testCtx.signErrorCount,
 		testCtx.fc)
 	test.AssertNotError(t, err, "Failed to create CA")
 
@@ -897,19 +841,19 @@ func TestIssueCertificateForPrecertificateDuplicateSerial(t *testing.T) {
 	errorca, err := NewCertificateAuthorityImpl(
 		errorsa,
 		testCtx.pa,
+		testCtx.ocsp,
 		testCtx.boulderIssuers,
 		nil,
 		testCtx.certExpiry,
 		testCtx.certBackdate,
 		testCtx.serialPrefix,
 		testCtx.maxNames,
-		testCtx.ocspLifetime,
 		testCtx.keyPolicy,
 		nil,
-		0,
-		time.Second,
 		testCtx.logger,
 		testCtx.stats,
+		testCtx.signatureCount,
+		testCtx.signErrorCount,
 		testCtx.fc)
 	test.AssertNotError(t, err, "Failed to create CA")
 
@@ -977,19 +921,19 @@ func TestPrecertOrphanQueue(t *testing.T) {
 	ca, err := NewCertificateAuthorityImpl(
 		qsa,
 		testCtx.pa,
+		testCtx.ocsp,
 		testCtx.boulderIssuers,
 		nil,
 		testCtx.certExpiry,
 		testCtx.certBackdate,
 		testCtx.serialPrefix,
 		testCtx.maxNames,
-		testCtx.ocspLifetime,
 		testCtx.keyPolicy,
 		orphanQueue,
-		0,
-		time.Second,
 		testCtx.logger,
 		testCtx.stats,
+		testCtx.signatureCount,
+		testCtx.signErrorCount,
 		testCtx.fc)
 	test.AssertNotError(t, err, "Failed to create CA")
 
@@ -1046,19 +990,19 @@ func TestOrphanQueue(t *testing.T) {
 	ca, err := NewCertificateAuthorityImpl(
 		qsa,
 		testCtx.pa,
+		testCtx.ocsp,
 		testCtx.boulderIssuers,
 		nil,
 		testCtx.certExpiry,
 		testCtx.certBackdate,
 		testCtx.serialPrefix,
 		testCtx.maxNames,
-		testCtx.ocspLifetime,
 		testCtx.keyPolicy,
 		orphanQueue,
-		0,
-		time.Second,
 		testCtx.logger,
 		testCtx.stats,
+		nil,
+		nil,
 		testCtx.fc)
 	test.AssertNotError(t, err, "Failed to create CA")
