+++ ca/ocsp_test.go
@@ -1,10 +1,14 @@
 package ca
 
 import (
+	"context"
+	"crypto/x509"
 	"encoding/hex"
 	"testing"
 	"time"
 
+	capb "github.com/letsencrypt/boulder/ca/proto"
+	"github.com/letsencrypt/boulder/core"
 	blog "github.com/letsencrypt/boulder/log"
 	"github.com/letsencrypt/boulder/metrics"
 	"github.com/letsencrypt/boulder/test"
@@ -18,6 +22,89 @@ func serial(t *testing.T) []byte {
 	}
 	return serial
 
+}
+
+func TestOCSP(t *testing.T) {
+	testCtx := setup(t)
+	ca, err := NewCertificateAuthorityImpl(
+		&mockSA{},
+		testCtx.pa,
+		testCtx.ocsp,
+		testCtx.boulderIssuers,
+		nil,
+		testCtx.certExpiry,
+		testCtx.certBackdate,
+		testCtx.serialPrefix,
+		testCtx.maxNames,
+		testCtx.keyPolicy,
+		nil,
+		testCtx.logger,
+		testCtx.stats,
+		testCtx.signatureCount,
+		testCtx.signErrorCount,
+		testCtx.fc)
+	test.AssertNotError(t, err, "Failed to create CA")
+	ocspi := testCtx.ocsp
+
+	// Issue a certificate from the RSA issuer caCert, then check OCSP comes from the same issuer.
+	rsaIssuerID := ca.issuers.byAlg[x509.RSA].ID()
+	rsaCertPB, err := ca.IssuePrecertificate(ctx, &capb.IssueCertificateRequest{Csr: CNandSANCSR, RegistrationID: arbitraryRegID})
+	test.AssertNotError(t, err, "Failed to issue certificate")
+	rsaCert, err := x509.ParseCertificate(rsaCertPB.DER)
+	test.AssertNotError(t, err, "Failed to parse rsaCert")
+	rsaOCSPPB, err := ocspi.GenerateOCSP(ctx, &capb.GenerateOCSPRequest{
+		Serial:   core.SerialToString(rsaCert.SerialNumber),
+		IssuerID: int64(rsaIssuerID),
+		Status:   string(core.OCSPStatusGood),
+	})
+	test.AssertNotError(t, err, "Failed to generate OCSP")
+	rsaOCSP, err := ocsp.ParseResponse(rsaOCSPPB.Response, caCert.Certificate)
+	test.AssertNotError(t, err, "Failed to parse / validate OCSP for rsaCert")
+	test.AssertEquals(t, rsaOCSP.Status, 0)
+	test.AssertEquals(t, rsaOCSP.RevocationReason, 0)
+	test.AssertEquals(t, rsaOCSP.SerialNumber.Cmp(rsaCert.SerialNumber), 0)
+
+	// Issue a certificate from the ECDSA issuer caCert2, then check OCSP comes from the same issuer.
+	ecdsaIssuerID := ca.issuers.byAlg[x509.ECDSA].ID()
+	ecdsaCertPB, err := ca.IssuePrecertificate(ctx, &capb.IssueCertificateRequest{Csr: ECDSACSR, RegistrationID: arbitraryRegID})
+	test.AssertNotError(t, err, "Failed to issue certificate")
+	ecdsaCert, err := x509.ParseCertificate(ecdsaCertPB.DER)
+	test.AssertNotError(t, err, "Failed to parse ecdsaCert")
+	ecdsaOCSPPB, err := ocspi.GenerateOCSP(ctx, &capb.GenerateOCSPRequest{
+		Serial:   core.SerialToString(ecdsaCert.SerialNumber),
+		IssuerID: int64(ecdsaIssuerID),
+		Status:   string(core.OCSPStatusGood),
+	})
+	test.AssertNotError(t, err, "Failed to generate OCSP")
+	ecdsaOCSP, err := ocsp.ParseResponse(ecdsaOCSPPB.Response, caCert2.Certificate)
+	test.AssertNotError(t, err, "Failed to parse / validate OCSP for ecdsaCert")
+	test.AssertEquals(t, ecdsaOCSP.Status, 0)
+	test.AssertEquals(t, ecdsaOCSP.RevocationReason, 0)
+	test.AssertEquals(t, ecdsaOCSP.SerialNumber.Cmp(ecdsaCert.SerialNumber), 0)
+
+	// GenerateOCSP with a bad IssuerID should fail.
+	_, err = ocspi.GenerateOCSP(context.Background(), &capb.GenerateOCSPRequest{
+		Serial:   core.SerialToString(rsaCert.SerialNumber),
+		IssuerID: int64(666),
+		Status:   string(core.OCSPStatusGood),
+	})
+	test.AssertError(t, err, "GenerateOCSP didn't fail with invalid IssuerID")
+
+	// GenerateOCSP with a bad Serial should fail.
+	_, err = ocspi.GenerateOCSP(context.Background(), &capb.GenerateOCSPRequest{
+		Serial:   "BADDECAF",
+		IssuerID: int64(rsaIssuerID),
+		Status:   string(core.OCSPStatusGood),
+	})
+	test.AssertError(t, err, "GenerateOCSP didn't fail with invalid Serial")
+
+	// GenerateOCSP with a valid-but-nonexistent Serial should *not* fail.
+	_, err = ocspi.GenerateOCSP(context.Background(), &capb.GenerateOCSPRequest{
+		Serial:   "03DEADBEEFBADDECAFFADEFACECAFE30",
+		IssuerID: int64(rsaIssuerID),
+		Status:   string(core.OCSPStatusGood),
+	})
+	test.AssertNotError(t, err, "GenerateOCSP failed with fake-but-valid Serial")
 }
 
 // Set up an ocspLogQueue with a very long period and a large maxLen,
